+++ providers/capa/standard/capa_test.go
@@ -20,7 +20,7 @@ var _ = Describe("Common tests", func() {
 
 	It("has all the control-plane nodes running", func() {
 		values := &application.ClusterValues{}
-		err := state.GetFramework().MC().GetHelmValues(state.GetCluster().Name, state.GetCluster().Namespace, values)
+		err := state.GetFramework().MC().GetHelmValues(state.GetCluster().Name, state.GetCluster().GetNamespace(), values)
 		Expect(err).NotTo(HaveOccurred())
 
 		wcClient, err := state.GetFramework().WC(state.GetCluster().Name)
@@ -34,7 +34,7 @@ var _ = Describe("Common tests", func() {
 
 	It("has all the worker nodes running", func() {
 		values := &application.ClusterValues{}
-		err := state.GetFramework().MC().GetHelmValues(state.GetCluster().Name, state.GetCluster().Namespace, values)
+		err := state.GetFramework().MC().GetHelmValues(state.GetCluster().Name, state.GetCluster().GetNamespace(), values)
 		Expect(err).NotTo(HaveOccurred())
 
 		wcClient, err := state.GetFramework().WC(state.GetCluster().Name)