+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Changed
+
+- Bumped clustertest to `v0.2.0` to use custom app values.
+
 ## [1.9.0] - 2023-08-21
 
 ### Added
@@ -30,7 +34,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Changed
 
-- Bumped clustertest to `v0.1.1` to fix `GITHUB_TOKEN` issue
+- Bumped clustertest to `v0.1.1` to fix `GITHUB_TOKEN` issue.
+- Bumped clustertest to `v0.2.0` to use custom app values.
 
 ### Added