+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Changed
 
 - Bumped clustertest to `v0.2.0` to use custom app values.
+- Refactored default app tests so they can log out current state during iterations
 
 ## [1.9.0] - 2023-08-21