+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Bumped clustertest to `v0.3.0` to be able to set the app `Organization`.
 - Bumped clustertest to `v0.2.0` to use custom app values.
+- Refactored default app tests so they can log out current state during iterations
 - Log out the non-running pods in test spec to show progress
 
 ## [1.9.0] - 2023-08-21